requirements.txt
numpy
scipy
scikit-learn
sphinx
mlflow
torch
pytest
google-cloud-storage
tqdm
sphinx_rtd_theme
ipykernel
pandas
Pillow
matplotlib
bs4
lxml
torchvision
click